OPR-163; Solve more feedback of PR
